Share storybook stories between the new and the previous implementation
