+++ src/medmain-modal/stories/index.js
@@ -1,3 +1,4 @@
+require('./get-started-stories');
 require('./modal-manager-stories');
 require('./context-api-stories');
 require('./component-stories');
